app/RuiaCollegeExample/requirement.txt
Flask-Bootstrap==2.3.2.2

Flask-Admin

Flask-SQLAlchemy

Flask-Security>=1.7.5

Flask-WTF==0.9.1

WTForms==1.0.4

itsdangerous==0.23

Flask-Login>=0.3.0
